Rip out some email endpoints
