+++ lib/api/email.js
@@ -1,0 +1,208 @@
+const routes = require('express').Router();
+const {build, send} = require('revolutionuc-emails');
+const {readOnlyAPIKeyValidator, apiKeyValidator, eventDeadlineValidator} = require('./mwValidators');
+
+const mailgunApiKey = process.env.MAILGUN_API_KEY;
+const mailgunDomain = process.env.MAILGUN_DOMAIN;
+
+/**
+ * Email route handlers
+ */
+
+/**
+ * This is used to send attendance confirmation emails to
+ * all registrants that:
+ *
+ *   - Have verified their email addresses
+ *   - Do not appear on the waitlist
+ *   - Have not previously received an attendance confirmation email
+ *
+ * It is intended to be called from the dashboard or manually.
+ *
+ * usage: /sendConfirmation?key=<apikeyhere>
+ */
+routes.get('/sendConfirmation', apiKeyValidator, eventDeadlineValidator, (request, response) => {
+  Registrant.find({emailVerified: true, waitList: false, receivedAttendanceConfirmationEmail: false}, (err, registrants) => {
+    if (err) {
+      console.error(err);
+      response.status(400).send('Sorry, something went wrong');
+    }
+
+    sendConfirmationEmails(request, registrants);
+    response.send({sent: true});
+  });
+});
+
+/**
+ * This is used to send attendance confirmation emails to
+ * all registrants that are sitting ducks. A sitting duck is
+ * a registrant that:
+ *
+ *   - Has previously received an attendance confirmation email
+ *   - But has not confirmed/denied their attendance
+ *
+ * It is intended to be called from the dashboard or manually.
+ * We use this to give people who haven't replied a bump so they
+ * can still confirm their attendance.
+ *
+ * usage: /sendConfirmationToSittingDucks?key=<apikeyhere>
+ */
+routes.get('/sendConfirmationToSittingDucks', apiKeyValidator, eventDeadlineValidator, (request, response) => {
+  Registrant.find({receivedAttendanceConfirmationEmail: true, attendanceConfirmed: 'NA'}, (err, registrants) => {
+    if (err) {
+      console.error(err);
+      response.status(400).send('Sorry, something went wrong');
+    }
+
+    sendConfirmationEmails(request, registrants);
+    response.send({sent: true});
+  });
+});
+
+/**
+ * This is used to send attendance confirmation emails to
+ * registrants that:
+ *
+ *   - Have registered after we hit our waitlist threshold
+ *   - Have verified their email address
+ *
+ * It is intended to be called from the dashboard or manually.
+ * We use this to give people who were put on the waitlist a chance
+ * to attend the event if space frees up.
+ *
+ * usage: /sendConfirmationToWaitlisted?key=<apikeyhere>
+ */
+routes.get('/sendConfirmationToWaitlisted', apiKeyValidator, eventDeadlineValidator, (request, response) => {
+  Registrant.find({emailVerified: true, waitList: true}, (err, registrants) => {
+    if (err) {
+      console.error(err);
+      response.status(400).send('Sorry, something went wrong');
+    }
+
+    Registrant.update({emailVerified: true, waitList: true}, {waitList: false}, {multi: true}, (err, numberAffected, rawResponse) => {
+      response.send(registrants);
+    });
+
+    sendConfirmationEmails(request, registrants, {offWaitlist: true});
+    response.json({"affected users": registrants.length});
+  });
+});
+
+/**
+ * This is used to send attendance confirmation emails to
+ * an individual registrant:
+ *
+ *   - Exec has decided to include regardless of any other condition
+ *
+ * It is intended to be called from the dashboard or manually.
+ * We use this to give priority to someone or to send an attendance
+ * confirmation email to someone who might not have received it.
+ * Use with caution :)
+ *
+ * usage: /sendConfirmationToIndividual?key=<apikeyhere>
+ */
+routes.get('/sendConfirmationToIndividual', apiKeyValidator, (request, response) => {
+  const userEmail = request.query.userEmail;
+
+  Registrant.find({email: userEmail}, (err, registrants) => {
+    if (err || !registrants.length) return response.status(400).send(`Either something went wrong, or ${userEmail} could not be found`);
+
+    Registrant.update({email: userEmail}, {waitList: false}, (err, numberAffected, rawResponse) => {
+      if (err) {
+        console.error(err);
+        response.status(400).send('Sorry, something went wrong');
+      }
+
+      sendConfirmationEmails(request, registrants);
+      response.json({"affected users": registrants.length});
+    }); // Registrant.update()
+  }); // Registrant.find()
+});
+
+/**
+ * This is used to send the infoEmail1 email to all registrants that:
+ *
+ *   - Have confirmed their email address
+ *
+ * It is intended to be called manually, or from the dashboard if that
+ * functionality gets implemented.
+ *
+ * usage: /sendInfoEmail1?key=<apikeyhere>
+ */
+routes.get('/sendInfoEmail1', apiKeyValidator, eventDeadlineValidator, (request, response) => {
+  Registrant.find({emailVerified: true}, (err, registrants) => {
+    if (err) {
+      console.error(err);
+      response.status(400).send('Sorry, something went wrong');
+    }
+
+    sendInfoEmail1(request, registrants);
+    response.json({"affected users": registrants.length});
+  });
+});
+
+/*
+ ***********************************************************
+ * Below are helper functions that some of the above
+ * endpoints call. For example, `/sendInfoEmail1` makes
+ * use of sendInfoEmail1(), while just about ever other
+ * email endpoint makes use of `sendConfirmationEmails()`
+ ***********************************************************
+ */
+
+async function sendConfirmationEmails(request, registrants, settings = {}) {
+  for (let registrant of registrants) {
+    // Build user hash
+    let cipher = crypto.createCipher(userCryptoAlgorithm, userCryptoPassword);
+    let encrypted = cipher.update(registrant.email, 'utf8', 'hex');
+    encrypted += cipher.final('hex');
+
+    // Build user confirmationUrl and nextTimeUrl
+    let confirmationUrl = `${request.protocol}://${request.get('host')}/registration/confirm?user=${encrypted}`;
+
+    // Build user emailData
+    let emailData = {
+      subject: 'Confirm Your Attendance',
+      shortDescription: 'Please confirm your attendance for RevolutionUC',
+      firstName: registrant.firstName,
+      yesConfirmationUrl: `${confirmationUrl}&confirm=YES`,
+      noConfirmationUrl: `${confirmationUrl}&confirm=NO`,
+    };
+
+    // Used to copy all enumerable own properties from `settings` to `emailData`, thus
+    // effectively merging the two objects. If `settings` and `emailData` share any of the
+    // same keys, the values associated with `settings` will be used, and will overwrite those
+    // in `emailData`
+    Object.assign(emailData, settings);
+
+    // Two async tasks
+    //   1. Send confirm attendance email
+    const html = await build('confirmAttendance', emailData);
+    send(mailgunApiKey, mailgunDomain, 'RevolutionUC <info@revolutionuc.com>', registrant.email, emailData.subject, html);
+
+    //   2. Update user in DB to have received the attendance confirmation email
+    Registrant.update({email: registrant.email}, {receivedAttendanceConfirmationEmail: true}, _ => {});
+  }
+}
+
+/**
+ * TODO(domfarolino): generalize this function for other "info"
+ * emails that we might send when the time comes.
+ */
+async function sendInfoEmail1(request, registrants) {
+  for (let registrant of registrants) {
+    // Build user emailData
+    let emailData = {
+      subject: 'RevolutionUC is Less Than 3 Weeks Away! 🙌 ',
+      shortDescription: "We're less than 3 weeks away from RevolutionUC Spring 2018! Here are some updates on the event.",
+      firstName: registrant.firstName
+    };
+
+    // One async task
+    //   1. Send infoEmail1
+    const html = await build('infoEmail1', emailData);
+    await send(mailgunApiKey, mailgunDomain, 'RevolutionUC <info@revolutionuc.com>', registrant.email, emailData.subject, html);
+  }
+}
+
+module.exports = routes;
