+++ lib/api/index.js
@@ -1,5 +1,6 @@
 const routes = require('express').Router();
 const registrationRoutes = require('./registration');
+const emailRoutes = require('./email');
 
 /**
  * CORS and Cache-handling stuff
@@ -31,7 +32,10 @@ routes.get('/', (request, response) => {
   response.json({api: true});
 });
 
-// Registration post handler
+// Registration endpoints
 routes.use('/registration', registrationRoutes);
+
+// Email endpoints
+routes.use('/email', emailRoutes);
 
 module.exports = routes;
