+++ lib/api/registration.js
@@ -263,52 +263,6 @@ routes.get('/verify', eventDeadlineValidator, (request, response) => {
   });
 });
 
-routes.get('/sendConfirmation', apiKeyValidator, eventDeadlineValidator, (request, response) => {
-  Registrant.find({emailVerified: true, waitList: false, receivedAttendanceConfirmationEmail: false}, (err, registrants) => {
-    if (err) return console.error(err);
-
-    sendConfirmationEmails(request, registrants);
-    response.send({sent: true});
-  });
-});
-
-routes.get('/sendConfirmationToSittingDucks', apiKeyValidator, eventDeadlineValidator, (request, response) => {
-  Registrant.find({receivedAttendanceConfirmationEmail: true, attendanceConfirmed: 'NA'}, (err, registrants) => {
-    if (err) return console.error(err);
-
-    sendConfirmationEmails(request, registrants);
-    response.send({sent: true});
-  });
-});
-
-routes.get('/sendConfirmationToWaitlisted', apiKeyValidator, eventDeadlineValidator, (request, response) => {
-  Registrant.find({emailVerified: true, waitList: true}, (err, registrants) => {
-    if (err) return console.error(err);
-
-    Registrant.update({emailVerified: true, waitList: true}, {waitList: false}, {multi: true}, (err, numberAffected, rawResponse) => {
-      response.send(registrants);
-    });
-
-    sendConfirmationEmails(request, registrants, {offWaitlist: true});
-  });
-});
-
-routes.get('/sendConfirmationIndividual', apiKeyValidator, (request, response) => {
-  const userEmail = request.query.userEmail;
-
-  Registrant.find({email: userEmail}, (err, registrants) => {
-    if (err || !registrants.length) return response.status(400).send(`Either something went wrong, or ${userEmail} could not be found`);
-
-    Registrant.update({email: userEmail}, {waitList: false}, (err, numberAffected, rawResponse) => {
-      if (err) return response.status(400).send('Sorry, something went wrong');
-
-      sendConfirmationEmails(request, registrants);
-    });
-
-    response.send({sent: true});
-  });
-});
-
 /**
  * This endpoint is intended to be called from the confirmation
  * email.
@@ -360,55 +314,6 @@ routes.get('/confirm', eventDeadlineValidator, (request, response) => {
   });
 });
 
-function sendConfirmationEmails(request, registrants, settings = {}) {
-  for (let registrant of registrants) {
-    // Build user hash
-    let cipher = crypto.createCipher(userCryptoAlgorithm, userCryptoPassword);
-    let encrypted = cipher.update(registrant.email, 'utf8', 'hex');
-    encrypted += cipher.final('hex');
-
-    // Build user confirmationUrl and nextTimeUrl
-    let confirmationUrl = `${request.protocol}://${request.get('host')}/registration/confirm?user=${encrypted}`;
-
-    // Build user emailData
-    let emailData = {
-      subject: 'Confirm Your Attendance',
-      shortDescription: 'Please confirm your attendance for RevolutionUC',
-      firstName: registrant.firstName,
-      yesConfirmationUrl: `${confirmationUrl}&confirm=YES`,
-      noConfirmationUrl: `${confirmationUrl}&confirm=NO`,
-    };
-
-    // Used to copy all enumerable own properties from `settings` to `emailData`, thus
-    // effectively merging the two objects. If `settings` and `emailData` share any of the
-    // same keys, the values associated with `settings` will be used, and will overwrite those
-    // in `emailData`
-    Object.assign(emailData, settings);
-
-    // Two async tasks
-    //   1. Send confirm attendance email
-    //   2. Update user in DB to have received the attendance confirmation email
-    email.build('confirmAttendance', emailData, 'RevolutionUC <info@revolutionuc.com>', registrant.email);
-    Registrant.update({email: registrant.email}, {receivedAttendanceConfirmationEmail: true}, _ => {});
-  }
-}
-
-async function sendInfoEmail1(request, registrants) {
-  for (let registrant of registrants) {
-    // Build user emailData
-    let emailData = {
-      subject: 'RevolutionUC is Less Than 3 Weeks Away! 🙌 ',
-      shortDescription: "We're less than 3 weeks away from RevolutionUC Spring 2018! Here are some updates on the event.",
-      firstName: registrant.firstName
-    };
-
-    // One async task
-    //   1. Send infoEmail1
-    const html = await build('infoEmail1', emailData);
-    await send(mailgunApiKey, mailgunDomain, 'RevolutionUC <info@revolutionuc.com>', registrant.email, emailData.subject, html);
-  }
-}
-
 /**
  * This is used to manually check a user in via
  * email address. It is intended to be used by the
@@ -426,21 +331,6 @@ routes.get('/checkIn', apiKeyValidator, (request, response) => {
   });
 });
 
-/**
- * This is used to send the email in revolutionuc-emails
- * denoted by name `infoEmail1`, to all registrants whose
- * email addresses are verified.
- */
-routes.get('/sendInfoEmail1', apiKeyValidator, eventDeadlineValidator, (request, response) => {
-  Registrant.find({emailVerified: true}, (err, registrants) => {
-    if (err) return response.status(400).send('Sorry, something went wrong');
-
-    sendInfoEmail1(request, registrants, {offWaitlist: true});
-    response.json({"sendInfoEmail1": "sent"});
-  });
-});
-
-
 routes.get('/all', readOnlyAPIKeyValidator, (request, response) => {
   Registrant.find((err, allRegistrants) => {
     if (err) return console.error(err);
@@ -448,11 +338,9 @@ routes.get('/all', readOnlyAPIKeyValidator, (request, response) => {
   });
 });
 
-/*
 routes.get('/clear', apiKeyValidator, (request, response) => {
   Registrant.remove({}, _ => {});
   response.json({removed: true});
 });
-*/
 
 module.exports = routes;
